refactor(targets): split TargetDetailV2.tsx into useTargetTonight, TonightPanel, LinkedItemsList, AliasesSection, NotesSection
C-15: TargetDetailV2 (950L) decomposed into:
- useTargetTonight.ts: tonight planner data derivation hook
- TonightPanel.tsx: altitude graph + stats + filter guidance
- LinkedItemsList.tsx: parameterized sessions/projects list (kills twin maps)
- AliasesSection.tsx: alias list + add form
- NotesSection.tsx: observing notes editor/viewer

TargetDetailV2 retains data loading + composition. Hook called unconditionally
with nullable coordinates for React rules-of-hooks compliance.
Behavior-preserving, typecheck green.
